#55 - Agrego parsers y valido input de usuario al cambiar dia de recordatorio
