+++ src/meals/tests/test_commands.py
@@ -500,3 +500,17 @@ martes 19 de abril _\\(id: {meal2.id}\\)_
 
         setting.refresh_from_db()
         setting.reminder_day == 0
+
+    @override_settings(CHAT_ID=1)
+    def test_change_reminder_uppercase_day(self, *args):
+        setting = CocaSettingsFactory(reminder_day=2)
+        context = get_mock_context(["luNes"])
+        update = get_mock_update()
+        change_reminder_handler(update, context)
+
+        update.message.reply_text.assert_called_once_with(
+            "Se actualizó el día del recordatorio al día lunes\\.",
+        )
+
+        setting.refresh_from_db()
+        setting.reminder_day == 0
